New Functionality For The RadioButtonGroupComponent
RadioButtonGroupComponent was lacking of the functionality that
SelectBoxComponent have with the setItems. Now the user can use setItems
member to feed the RadioButtonGroupComponent with and array or an object
and then the RadioButtonGroupComponent will create the radio
buttons.

Fix bug in RadioButtonGroupComponent there isn't any UserData in
generated radio buttons.