Minor refactoring in container component
Summary:
Container component now utilize getContainerWidget in
add/contains/remove methods in order sub-classes to be able use diffrent
composite for the main widget and diffrent composite for the children
container. So in the container widget is where the children will be
added, and in the main widget there would be space for controls or other
stuff.