Try-catch block for event handlers
Event listeners are typically not protected against unexpected errors.
This has the side effect that, when one listener throws an exception,
subsequent listeners are never called. In the context of Qookery, this
is unwanted, since XML authors are permitted to fail anyway they please
without having to mind about potential side effects in code they do not
control (e.g. in <script event="appear"> of windows).